wifi: Native registerForNotifications to pass the "default" instance name.
This helps to co-locate with other HAL's (wigig). Without which, the
notifications from wigig clients shall also be received and thus resulting
in turning down the native instances too.
Thus, pass the instance name as "default" as the native servers are
also registered with "default". This would get the notifications only
from the matching (native) peers.

Change-Id: I0ba6cd099985fd3763f0dec7c2a6ce1c50ec8b5c
CRs-Fixed: 2412094